vp-license/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
PyJWT==2.9.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
slowapi==0.1.9
cryptography==43.0.1
python-multipart==0.0.12
